// src/app/types/LiveReels.ts
export interface ActivityOnLiveComment {
    comment : boolean;
    comment_cotent:string;
    current_like:number;
    first_name:string;
    full_name:string;
    last_name:string;
    like:boolean;
    profile_pic:string;
    total_comments:number;
    user_id:number;
    user_name:string;
}

// Activity on Live Like Res =============================== 
export interface ActivityOnLiveLike {
    comment : boolean;
    comment_cotent:string;
    current_like:number;
    first_name:string;
    full_name:string;
    last_name:string;
    like:boolean;
    profile_pic:string;
    total_comments:number;
    user_id:number;
    user_name:string;
}